Refactor the task and extension configuration in the plugin
Now uses PropertyLookup from atlas-unity
