--- src/main/groovy/wooga/gradle/xcodebuild/internal/PropertyLookup.groovy
@@ -1,29 +1,0 @@
-/*
- * Copyright 2018-2020 Wooga GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package wooga.gradle.xcodebuild.internal
-
-class PropertyLookup {
-    final String env
-    final String property
-    final String defaultValue
-
-    PropertyLookup(String env, String property, String defaultValue) {
-        this.env = env
-        this.property = property
-        this.defaultValue = defaultValue
-    }
-}
